requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.9
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
openapi-codec==1.3.2
pytz==2021.3
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
